mpdscribble: Initial commit
My personal PKGBUILD for mpdscribble. It uses soup instead of curl
because of a incompatibility introduced by curl 7.21.1. This package
don't have a rcd script as it is intended to be executed by a normal
user. Also, a little patch is applied to the Makefile.in which prevents
the config file being installed as f=mpdscribble.conf instead of
mpdscribble.conf. Again, as this package don't aims to be executed by a
rcd script, the config file is just copied to the docs dir for
reference.