emacs-pov-mode: Fix misleading message/Full review
The message printed in post-install() was missing some escape characters
that led to a wrong printed message possibly confusing a user that is
new to Emacs. This change resulted in a bump to the package release.

Besides, a full review was done. Change from md5sums to sha256sums (big
enough, but not too big like sha512sums). Use of double quotes just were
it is really needed (variables in the string) or for better visual
appearance. Use of braces just when it helps visualize the limits of the
variable name or to use some kind of bash feature. Cleanup.