minbif: Full review
Change from md5sums to sha256sums (big enough, but not too big like
sha512sums). Use of double quotes just were it is really needed
(variables in the string) or for better visual appearance. Use of braces
just when it helps visualize the limits of the variable name or to use
some kind of bash feature. Cleanup.

There was a discussion about mounting /var/run as a tmpfs. As a result,
the creation of /var/run/minbif has to be handled by the rcd script and
not by the package anymore. Changes were made to accomplish this.