mpdscribble: New upstream release v0.22
This commit adds the rcd script that was missing in the old package.
There are some differences between this and the package from community:
instead of "moving" the config file to /etc/mpdscribble.conf.example
(mode 644), it maintains the normal config file with the right
permissions (mode 600 - the file can contain passwords) and install a
copy of the file in /usr/share/doc/mpdscribble/mpdscribble.conf (mode
644) intended for per-user usage; change back to curl (the default)
instead of libsoup.