vim-renamer: Full review
Change from md5sums to sha256sums (big enough, but not too big like
sha512sums). Use of double quotes just were it is really needed
(variables in the string) or for better visual appearance. Use of braces
just when it helps visualize the limits of the variable name or to use
some kind of bash feature. Bump the package release to be on par with
the AUR package. License changed to "unknown". Cleanup.