bitlbee: Initial commit
My personal PKGBUILD for bitlbee. I thought it would be a good idea to
make a bitlbee package using the PKGBUILD from minbif as a base. So here
it is!

This package uses gnutls instead of openssl used by the package in
extra. Besides, it doesn't print the error message presented by the
package in the repos when removing the package (not a big problem).