rxvt-unicode: Initial commit
My personal PKGBUILD for rxvt-unicode. This is a mix of the package from
the extra repo and the rxvt-unicode-patched from AUR. It just adds a
modified version of the font patch to "fix" an issue with the Xft font
rendering. It doesn't include the layout gaps patch.

The original font patch just "fix" the width of the fonts. But in fact,
if you compare the rendering between xterm and urxvt you'll see that the
line spacing is much greater in urxvt. This annoyed me for some time, so
I began to investigate. First I tried the easy way, just looking at the
modifications done by the patch and doing the same thing for the height.
It made sense at first thought, but nothing changed. Until I decided to
take a look in the source code of xterm to see how it handles this. I
finally found the "problem" and made the changes. I don't really know if
this is the real "fix" for this, but it's working so far.