bsnes: New upstream release v088
Patch was removed in favor of the new bash script to launch bsnes. The
script will try to copy some files needed by bsnes (if they don't exist
yet) in your config folder (~/.config/bsnes). The script won't check for
updated files, which is a failure that I'll _try_ to address later.

Some major rewriting and cleanup was made to the PKGBUILD too.